Finished CUDA matrix multi. Over 60GFLOPS, when using blocked-multi style. Multiprocessor's shared memory is 16KB, so blocking is not very powerful.

+++ cpp/lib_calvin/cuda/matrix_multi.h
@@ -1,5 +1,5 @@
 #include "cuda_runtime.h"
 
-cudaError_t matrixMulti(double *a, double *b, double *target, size_t size);
+cudaError_t matrixMulti(int *a, int *b, int *target, size_t size);
 
 void matrixMultiTest();
